Disintegrate Cardano.Wallet.Launch module
